pyproject.toml
[project]
name = "repo_template"
version = "0.0.1"
description = "Repo Template"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "TT", email = "TT@aol.com"},
]
license = {file = "LICENSE"}

dependencies = [
  "typer-cli", # 0.12.0
  "wasabi",
  "termcolor",
  "pyfiglet",
  # "colorama",
  "terminology",
]

[tool.setuptools.packages.find]
where = ["."]
namespaces = false

[project.scripts]
repo_template = "repo_template.cli:main"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# This field adds keywords for your project which will appear on the
# project page. What does your project relate to?
#
# Note that this is a list of additional keywords, separated
# by commas, to be used to assist searching for the distribution in a
# larger catalog.
# keywords = ["sample", "setuptools", "development"]

# [project.urls]
# "Homepage" = "https://github.com/pypa/sampleproject"
# "Bug Reports" = "https://github.com/pypa/sampleproject/issues"
# "Funding" = "https://donate.pypi.org"
# "Say Thanks!" = "http://saythanks.io/to/example"
# "Source" = "https://github.com/pypa/sampleproject/"
